% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSImetadata.R
\docType{data}
\name{HSImetadata}
\alias{HSImetadata}
\title{Habitat suitability index (HSI) model metadata}
\format{
A data frame with 351 rows and 85 variables:

\describe{
  \item{model}{Model name}
  \item{submodel}{Model specifications}
  \item{species}{Scientific nomenclature of modeled taxa}
  \item{geography}{Geographic range of organism}
  \item{ecosystem}{Type of habitat}
  \item{documentation}{Citation of original model}
  \item{note}{Conditions under which model may be applied}
  \item{website}{Link to individual model source}
  \item{SIV1}{Suitability index values for each organism specific condition}
  \item{SIV1B}{Suitability index values for each organism specific condition}
  \item{SIV2}{Suitability index values for each organism specific condition}
  \item{SIV2B}{Suitability index values for each organism specific condition}
  \item{SIV3}{Suitability index values for each organism specific condition}
  \item{SIV3B}{Suitability index values for each organism specific condition}
  \item{SIV4}{Suitability index values for each organism specific condition}
  \item{SIV4B}{Suitability index values for each organism specific condition}
  \item{SIV5}{Suitability index values for each organism specific condition}
  \item{SIV5B}{Suitability index values for each organism specific condition}
  \item{SIV6}{Suitability index values for each organism specific condition}
  \item{SIV6B}{Suitability index values for each organism specific condition}
  \item{SIV7}{Suitability index values for each organism specific condition}
  \item{SIV7B}{Suitability index values for each organism specific condition}
  \item{SIV8}{Suitability index values for each organism specific condition}
  \item{SIV8B}{Suitability index values for each organism specific condition}
  \item{SIV9}{Suitability index values for each organism specific condition}
  \item{SIV10}{Suitability index values for each organism specific condition}
  \item{SIV11}{Suitability index values for each organism specific condition}
  \item{SIV12}{Suitability index values for each organism specific condition}
  \item{SIV13}{Suitability index values for each organism specific condition}
  \item{SIV14}{Suitability index values for each organism specific condition}
  \item{SIV15}{Suitability index values for each organism specific condition}
  \item{SIV15B}{Suitability index values for each organism specific condition}
  \item{SIV16}{Suitability index values for each organism specific condition}
  \item{SIV16B}{Suitability index values for each organism specific condition}
  \item{SIV17}{Suitability index values for each organism specific condition}
  \item{SIV18}{Suitability index values for each organism specific condition}
  \item{SIV19}{Suitability index values for each organism specific condition}
  \item{SIV20}{Suitability index values for each organism specific condition}
  \item{SIV21}{Suitability index values for each organism specific condition}
  \item{SIV22}{Suitability index values for each organism specific condition}
  \item{CF}{Food component equation}
  \item{CRF}{Food/reproduction component equation}
  \item{CRN}{Roosting-nesting component equation}
  \item{CC}{Cover component equation}
  \item{CCRO}{Cover roosting component equation}
  \item{CCRF}{Cover-reproduction-food component equation}
  \item{CCF}{Cover-food component equation}
  \item{CCSF}{Cover-food shrub component equation}
  \item{CCHF}{Cover-food herbaceous component equation}
  \item{CWF}{Winter food component}
  \item{CSF}{Summer food component}
  \item{CFF}{Fall food component}
  \item{CW}{Water component}
  \item{CCB}{Cover breeding component}
  \item{CB}{Brood component}
  \item{CN}{Nest component}
  \item{CNBC}{Nest-brood cover component}
  \item{CCN}{Cover nesting component}
  \item{CP}{Pair habitat component}
  \item{CWQ}{Water quality component}
  \item{CR}{Reproduction component}
  \item{CCR}{Cover reproduction component}
  \item{CD}{Disturbance component}
  \item{COT}{Other component}
  \item{CL}{Larval component}
  \item{CEL}{Embryo and larval component}
  \item{CE}{Embryo component}
  \item{CJ}{Juvenile component}
  \item{CFr}{Fry component}
  \item{CS}{Spawning component}
  \item{CA}{Adult component}
  \item{CI}{Island component}
  \item{CIN}{Interspersion component}
  \item{CNI}{Non-island component}
  \item{CWFC}{Winter cover food component}
  \item{CFBS}{Summer food brood component}
  \item{CFSWF}{Fall spring winter food component}
  \item{CSPF}{Spring food component}
  \item{CWC}{Winter cover component}
  \item{CCFS}{Fall to spring cover component}
  \item{CSS}{Substrate-suspended solids component}
  \item{CT}{Topography component}
  \item{CTe}{Temperature component}
  \item{CJA}{Juvenile adult component}
  \item{Eqtn}{HSI overarching model equation in R syntax}
}
}
\source{
\url{https://pubs.usgs.gov/}
}
\usage{
HSImetadata
}
\description{
Metadata for 351 U.S. Fish and Wildlife Service Habitat suitability index (HSI) models
}
\keyword{datasets}
